adds circuit id components
