+++ tgstation.dme
@@ -4053,6 +4053,10 @@
 #include "code\modules\wiremod\components\bci\hud\counter_overlay.dm"
 #include "code\modules\wiremod\components\bci\hud\object_overlay.dm"
 #include "code\modules\wiremod\components\bci\hud\target_intercept.dm"
+#include "code\modules\wiremod\components\id\access_checker.dm"
+#include "code\modules\wiremod\components\id\access_reader.dm"
+#include "code\modules\wiremod\components\id\getter.dm"
+#include "code\modules\wiremod\components\id\info_reader.dm"
 #include "code\modules\wiremod\components\list\assoc_literal.dm"
 #include "code\modules\wiremod\components\list\concat.dm"
 #include "code\modules\wiremod\components\list\filter.dm"
